Clean up and add collapsing comment feature
Clean up post trpc route
Simplify normalizeReplies function to use a plain JS object instead of a Map
Add type to force subredditName param of fetchSubredditPosts to be without /r/ or r/
Clean up fetchSubredditPosts and add /r/ prefix after reddit base url
Add collapsing comment button and functionality

Update remeasure function of ListVirtualizerContext to not use the native measure function of the virtualizer
because it doesn't do what you think it does. Replace it with a scrollToOffset where the offset to scroll to
is the current scrollY + 1.